+++ src/main/java/com/example/demo/dto/review/ReviewSearchRequest.java
@@ -1,4 +1,6 @@
-package com.example.demo.dto;
+package com.example.demo.dto.review;
+
+import com.example.demo.dto.PagedRequest;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
