+++ src/main/java/com/example/demo/mapper/CafeMapper.java
@@ -9,7 +9,6 @@ import com.example.demo.domain.cafe.CafeImpl;
 import com.example.demo.domain.cafe.CafeSearchCondition;
 import com.example.demo.domain.cafe.OpenChecker;
 import com.example.demo.dto.CanMakeStudyOnceResponse;
-import com.example.demo.dto.ReviewResponse;
 import com.example.demo.dto.StudyOnceForCafeResponse;
 import com.example.demo.dto.cafe.BusinessHourResponse;
 import com.example.demo.dto.cafe.CafeBasicInfoResponse;
@@ -17,6 +16,7 @@ import com.example.demo.dto.cafe.CafeResponse;
 import com.example.demo.dto.cafe.CafeSearchRequest;
 import com.example.demo.dto.cafe.CafeSearchResponse;
 import com.example.demo.dto.cafe.SnsResponse;
+import com.example.demo.dto.review.ReviewResponse;
 
 public class CafeMapper {
