+++ src/main/java/com/example/demo/mapper/ReviewMapper.java
@@ -3,11 +3,11 @@ package com.example.demo.mapper;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.example.demo.domain.ReviewImpl;
 import com.example.demo.domain.member.MemberImpl;
-import com.example.demo.dto.ReviewResponse;
-import com.example.demo.dto.ReviewSearchResponse;
+import com.example.demo.domain.review.ReviewImpl;
 import com.example.demo.dto.WriterResponse;
+import com.example.demo.dto.review.ReviewResponse;
+import com.example.demo.dto.review.ReviewSearchResponse;
 
 public class ReviewMapper {
