+++ src/main/java/com/example/demo/repository/review/ReviewRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.repository;
+package com.example.demo.repository.review;
 
 import java.util.List;
 
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.example.demo.domain.ReviewImpl;
+import com.example.demo.domain.review.ReviewImpl;
 
 public interface ReviewRepository extends JpaRepository<ReviewImpl, Long> {
