+++ src/main/java/com/example/demo/service/review/ReviewQueryService.java
@@ -1,9 +1,9 @@
-package com.example.demo.service;
+package com.example.demo.service.review;
 
 import com.example.demo.dto.PagedResponse;
-import com.example.demo.dto.ReviewResponse;
-import com.example.demo.dto.ReviewSearchRequest;
-import com.example.demo.dto.ReviewSearchResponse;
+import com.example.demo.dto.review.ReviewResponse;
+import com.example.demo.dto.review.ReviewSearchRequest;
+import com.example.demo.dto.review.ReviewSearchResponse;
 
 public interface ReviewQueryService {
