+++ src/test/java/com/example/demo/repository/ReviewRepositoryTest.java
@@ -16,14 +16,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.config.QueryDslConfig;
 import com.example.demo.config.TestConfig;
-import com.example.demo.domain.ReviewImpl;
 import com.example.demo.domain.cafe.CafeImpl;
 import com.example.demo.domain.member.MemberImpl;
 import com.example.demo.domain.member.ThumbnailImage;
+import com.example.demo.domain.review.ReviewImpl;
 import com.example.demo.helper.CafePersistHelper;
 import com.example.demo.helper.MemberPersistHelper;
 import com.example.demo.helper.ReviewPersistHelper;
 import com.example.demo.helper.ThumbnailImagePersistHelper;
+import com.example.demo.repository.review.ReviewRepository;
 import com.example.demo.util.PageRequestCustom;
 
 @DataJpaTest
